update test bases so that only tearsdown when entity created in setup exists

+++ src/test/java/com/redhat/qe/test/rest/ClusterTestBase.java
@@ -25,8 +25,9 @@ public abstract class ClusterTestBase extends RestTestBase {
 	}
 
 	@After
-	public void cleanupCluster() {
-		getClusterRepository().destroy(cluster);
+	public void cleanup() {
+		if(cluster != null && cluster.getId() != null)
+			getClusterRepository().destroy(cluster);
 	}
 
 }
